Save JWT-created user to DB
Without committing this session the User object we created was not
saved, meaning that we would create a new on every time (which wasn't
what I intended) - this was visible by seeing the "Creating airflow
user details for ... from JWT" appear once per request in the logs.

This fixes that problem, and also means the user can edit their name now

Since we are now actually explicitly committing in our code we need to
create an extra nested transaction in our test harness. I have also
changed to a simpler method of detecting an unexpected commit -
listening to the event on the db connection.