Reload JWT cert signing path when file is modified (#8)
To not hit the disk too much we don't look at the JWT token for requests
to /static/ (CSS, JS, etc). This _might_ mean that some workers have the
old one and reject a request for a static asset, but those should be
fixed within 10s by the request for the /health endpoint.