Add current user backend (#17)
This adds a new authentication backend that exclusively checks the current_user value.
This works when the Astro platform uses the `before_request` method in the security
manager to decode a JWT and populate the `current_user` value.

To specify this backend in production, add this line to your dockerfile.

`ENV AIRFLOW__API__AUTH_BACKEND=astronomer.flask_appbuilder.current_user_backend`